Refactor : 함수정의와 main 분리, 17번, 31번 수정 - LottoNumber.swift 생성, 함수정의와 실행구문 분리 - 17번 줄 return 값 이름변경 randomNumber -> lottoNumber - 31번 수식표현 변경  "mergeSameNumber = mergeSameNumber +" -> "mergeSameNumber +="
